Move api base classes into common module
